Make widgets independent from each other
As it is, changing one of the selects affects all widgets on the page.
This is obviously undesirable. We therefore load the actual widgets
with XHRs to make them independent. This has the additional benefit of
an improved UX, and also that the widgets will not be displayed at all
if the browser doesn't support JavaScript.