Drastically increase default resize dimensions
Scaling is not meant to create thumbnails, but rather to allow big
images to be down scaled prior to upload.  This is mostly relevant for
images taken with contemporary digital cameras, which may have 10M or
even more pixels.  Our large size is now roughly 5M, while the medium
size is about 1M.

We also add some words to the user manual about the resizing.